関数まとめ
とりあえずまとめた
