--- src/nilCheck.go
@@ -1,13 +1,0 @@
-package src
-
-import (
-	"github.com/bwmarrin/discordgo"
-	"log"
-)
-
-func nilCheck(Err *discordgo.Message) {
-	if Err != nil {
-		log.Println("Error: ", Err)
-		return
-	}
-}
