+++ src/tools.go
@@ -2,6 +2,7 @@ package src
 
 import (
 	"github.com/bwmarrin/discordgo"
+	"log"
 	"strings"
 )
 
@@ -20,4 +21,14 @@ func MessageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
 	nilCheck(Err)
 }
 
+func BootNotify(s *discordgo.Session, m *discordgo.Ready) {
+	Err, _ := s.ChannelMessageSend("699941274484080660", "BootBot! <@!622077711309078529>")
+	nilCheck(Err)
+}
 
+func nilCheck(Err *discordgo.Message) {
+	if Err != nil {
+		log.Println("Error: ", Err)
+		return
+	}
+}
